requirements.txt
streamlit
mistralai
python-dotenv
pyttsx3
